README.md: Use reference-style links for lesson-template
Full URLs are for browsers ;), and it's nice to give some
human-oriented text explaining what you're linking to instead.  In
this case, readers don't need an up-front reminder that the
lesson-template project is hosted on GitHub, or that we suggest they
access it over GitHub's prefered HTTPS protocol.  They just need to
know that we're pointing them at the lesson-template project.

I've also switched from an inline link to a reference-style link [1],
because it get's the link target out of the text, which makes for
easier reading (no URL clutter) and future updates (because most
version control systems use line-based diffs and reference-style links
put the link text and target on different lines).

[1]; http://daringfireball.net/projects/markdown/syntax#link